+++ Assets/UniGLTF/Editor/UniGLTF/ExportDialog/ExportDialogBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 292cdc4b7dd3b6d47a1215ad528f50c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
